% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa.R
\name{sa_balance}
\alias{sa_balance}
\title{Fetch balance sheet data}
\usage{
sa_balance(ticker)
}
\arguments{
\item{ticker}{Stock ticker}
}
\value{
tibble: ticker, metric, period, fiscal_year, is_ttm, value
}
\description{
Fetch balance sheet data
}
