% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa.R
\name{sa_balance_quarterly}
\alias{sa_balance_quarterly}
\title{Fetch quarterly balance sheet}
\usage{
sa_balance_quarterly(ticker)
}
\arguments{
\item{ticker}{Stock ticker}
}
\value{
tibble: ticker, metric, period, fiscal_year, is_ttm, value
}
\description{
Fetch quarterly balance sheet
}
