% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa.R
\name{sa_cashflow}
\alias{sa_cashflow}
\title{Fetch cash flow statement}
\usage{
sa_cashflow(ticker)
}
\arguments{
\item{ticker}{Stock ticker}
}
\value{
tibble: ticker, metric, period, fiscal_year, is_ttm, value
}
\description{
Fetch cash flow statement
}
