% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa.R
\name{sa_cashflow_quarterly}
\alias{sa_cashflow_quarterly}
\title{Fetch quarterly cash flow statement}
\usage{
sa_cashflow_quarterly(ticker)
}
\arguments{
\item{ticker}{Stock ticker}
}
\value{
tibble: ticker, metric, period, fiscal_year, is_ttm, value
}
\description{
Fetch quarterly cash flow statement
}
