% Generated by build script
\name{sa_income_quarterly}
\alias{sa_income_quarterly}
\title{Fetch quarterly income statement}
\usage{
sa_income_quarterly()
}
\arguments{
\item{ticker}{Stock ticker}
}
\value{tibble: ticker, metric, period, fiscal_year, is_ttm, value}
\description{Fetch quarterly income statement}
