% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa.R
\name{sa_ratios}
\alias{sa_ratios}
\title{Fetch financial ratios}
\usage{
sa_ratios(ticker)
}
\arguments{
\item{ticker}{Stock ticker}
}
\value{
tibble: ticker, metric, period, fiscal_year, is_ttm, value
}
\description{
PE, PB, PS, dividend yield, margins, returns, etc.
}
