% Generated by build script
\name{sa_balance_quarterly}
\alias{sa_balance_quarterly}
\title{Fetch quarterly balance sheet}
\usage{
sa_balance_quarterly()
}
\arguments{
\item{ticker}{Stock ticker}
}
\value{tibble: ticker, metric, period, fiscal_year, is_ttm, value}
\description{Fetch quarterly balance sheet}
