% Generated by build script
\name{sa_cashflow_quarterly}
\alias{sa_cashflow_quarterly}
\title{Fetch quarterly cash flow statement}
\usage{
sa_cashflow_quarterly()
}
\arguments{
\item{ticker}{Stock ticker}
}
\value{tibble: ticker, metric, period, fiscal_year, is_ttm, value}
\description{Fetch quarterly cash flow statement}
