% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa.R
\name{sa_income_quarterly}
\alias{sa_income_quarterly}
\title{Fetch quarterly income statement}
\usage{
sa_income_quarterly(ticker)
}
\arguments{
\item{ticker}{Stock ticker}
}
\value{
tibble: ticker, metric, period, fiscal_year, is_ttm, value
}
\description{
Fetch quarterly income statement
}
